ref(seer): Split resource embed tests
