--- static/app/components/seer/markdown/embeds/components/resourceEmbeds.spec.tsx
@@ -1,485 +1,0 @@
-import {DashboardFixture} from 'sentry-fixture/dashboard';
-import {EventsStatsFixture} from 'sentry-fixture/events';
-import {WidgetFixture} from 'sentry-fixture/widget';
-import {WidgetQueryFixture} from 'sentry-fixture/widgetQuery';
-
-import {render, screen, userEvent, waitFor} from 'sentry-test/reactTestingLibrary';
-
-import {SeerMarkdown} from 'sentry/components/seer/markdown';
-import {ConfigStore} from 'sentry/stores/configStore';
-import type {Config} from 'sentry/types/system';
-import {DisplayType, WidgetType} from 'sentry/views/dashboards/types';
-
-interface RenderEmbedOptions {
-  data: Record<string, unknown>;
-  name: string;
-  level?: 'block' | 'inline';
-}
-
-function renderEmbed({name, data, level = 'block'}: RenderEmbedOptions) {
-  const tag = `{% ${name} %}${JSON.stringify(data)}{% /${name} %}`;
-  const raw = level === 'inline' ? `text ${tag} text` : tag;
-  return render(<SeerMarkdown raw={raw} />);
-}
-
-function hrefFor(name: string, label: string, data: Record<string, unknown>) {
-  renderEmbed({name, data, level: 'inline'});
-  return screen.getByRole('link', {name: label}).getAttribute('href') ?? '';
-}
-
-describe('Seer resource embeds', () => {
-  let initialConfig: Config;
-
-  beforeEach(() => {
-    initialConfig = ConfigStore.getState();
-  });
-
-  afterEach(() => {
-    ConfigStore.loadInitialData(initialConfig);
-  });
-
-  it('links a dashboard title to the dashboard in the current organization', async () => {
-    const {router} = renderEmbed({
-      name: 'dashboard',
-      level: 'inline',
-      data: {
-        id: '123',
-        title: 'Application health',
-      },
-    });
-
-    await userEvent.click(screen.getByRole('link', {name: 'Application health'}));
-
-    expect(router.location.pathname).toBe('/organizations/org-slug/dashboard/123/');
-  });
-
-  it('uses a dashboard fallback label and normalizes customer-domain links', () => {
-    ConfigStore.set('customerDomain', {
-      subdomain: 'org-slug',
-      organizationUrl: 'https://org-slug.sentry.io',
-      sentryUrl: 'https://sentry.io',
-    });
-
-    renderEmbed({name: 'dashboard', data: {id: '456'}, level: 'inline'});
-
-    expect(screen.getByRole('link', {name: 'Dashboard 456'})).toHaveAttribute(
-      'href',
-      '/dashboard/456/'
-    );
-  });
-
-  it('renders a live preview of the first four dashboard widgets', async () => {
-    const widgets = ['Errors', 'Latency', 'Users', 'Throughput', 'Slow spans'].map(
-      (title, index) =>
-        WidgetFixture({
-          id: String(index + 1),
-          title,
-          displayType: index === 0 ? DisplayType.LINE : DisplayType.TEXT,
-          widgetType: index === 0 ? WidgetType.ERRORS : undefined,
-          description: `${title} details`,
-          limit: index === 0 ? 20 : undefined,
-        })
-    );
-    const dashboardRequest = MockApiClient.addMockResponse({
-      url: '/organizations/org-slug/dashboards/123/',
-      body: DashboardFixture(widgets, {
-        environment: ['production'],
-        id: '123',
-        projects: [1],
-        title: 'Application health',
-        utc: true,
-      }),
-    });
-    const widgetDataRequest = MockApiClient.addMockResponse({
-      url: '/organizations/org-slug/events-stats/',
-      body: EventsStatsFixture(),
-    });
-    MockApiClient.addMockResponse({
-      url: '/organizations/org-slug/releases/stats/',
-      body: [],
-    });
-
-    const {unmount} = renderEmbed({
-      name: 'dashboard',
-      data: {id: '123'},
-    });
-
-    expect(
-      await screen.findByRole('link', {name: 'Application health'}, {timeout: 5_000})
-    ).toBeInTheDocument();
-    expect(screen.getByText('Errors')).toBeInTheDocument();
-    expect(screen.getByText('Latency')).toBeInTheDocument();
-    expect(screen.getByText('Users')).toBeInTheDocument();
-    expect(screen.getByText('Throughput')).toBeInTheDocument();
-    expect(screen.queryByText('Slow spans')).not.toBeInTheDocument();
-    expect(screen.getByText('5 widgets')).toBeInTheDocument();
-    expect(screen.getByRole('link', {name: 'View 1 more widget'})).toHaveAttribute(
-      'href',
-      '/organizations/org-slug/dashboard/123/'
-    );
-    expect(dashboardRequest).toHaveBeenCalled();
-    await waitFor(() =>
-      expect(widgetDataRequest).toHaveBeenCalledWith(
-        '/organizations/org-slug/events-stats/',
-        expect.objectContaining({
-          query: expect.objectContaining({
-            environment: ['production'],
-            interval: '10m',
-            project: [1],
-            statsPeriod: '24h',
-            topEvents: 5,
-          }),
-        })
-      )
-    );
-    unmount();
-  });
-
-  it('keeps dashboard legend interactions local to the embed', async () => {
-    const widget = WidgetFixture({
-      id: '1',
-      title: 'Errors',
-      displayType: DisplayType.LINE,
-      widgetType: WidgetType.ERRORS,
-      queries: [
-        WidgetQueryFixture({
-          name: 'Current',
-          conditions: 'release:current',
-          fields: ['count()'],
-          aggregates: ['count()'],
-          columns: [],
-        }),
-        WidgetQueryFixture({
-          name: 'Previous',
-          conditions: 'release:previous',
-          fields: ['count()'],
-          aggregates: ['count()'],
-          columns: [],
-        }),
-      ],
-    });
-    MockApiClient.addMockResponse({
-      url: '/organizations/org-slug/dashboards/123/',
-      body: DashboardFixture([widget], {id: '123', title: 'Application health'}),
-    });
-    MockApiClient.addMockResponse({
-      url: '/organizations/org-slug/events-stats/',
-      body: EventsStatsFixture(),
-    });
-    MockApiClient.addMockResponse({
-      url: '/organizations/org-slug/releases/stats/',
-      body: [],
-    });
-
-    const {router, unmount} = renderEmbed({name: 'dashboard', data: {id: '123'}});
-    await userEvent.click(await screen.findByRole('button', {name: '+2 more'}));
-    const option = await screen.findByRole('option', {name: /Current : count\(\)/});
-
-    expect(option).toHaveAttribute('aria-selected', 'true');
-
-    await userEvent.click(option);
-
-    expect(option).toHaveAttribute('aria-selected', 'false');
-    expect(router.location.query.unselectedSeries).toBeUndefined();
-    unmount();
-  });
-
-  it('shows an error notice when dashboard details fail to load', async () => {
-    MockApiClient.addMockResponse({
-      url: '/organizations/org-slug/dashboards/123/',
-      statusCode: 500,
-    });
-
-    renderEmbed({name: 'dashboard', data: {id: '123'}});
-
-    expect(await screen.findByRole('alert')).toHaveTextContent(
-      'Unable to load dashboard details.'
-    );
-  });
-
-  it('links a replay to the relevant event timestamp (inline)', async () => {
-    const {router} = renderEmbed({
-      name: 'replay',
-      data: {
-        id: '4c1f2e3d1234567890',
-        eventTimestamp: '2026-08-25T16:37:12Z',
-      },
-      level: 'inline',
-    });
-
-    await userEvent.click(screen.getByRole('link', {name: 'Replay 4c1f2e3d'}));
-
-    expect(router.location.pathname).toBe(
-      '/organizations/org-slug/explore/replays/4c1f2e3d1234567890/'
-    );
-    expect(router.location.query.event_t).toBe('2026-08-25T16:37:12Z');
-  });
-
-  it('links a replay without a timestamp to the beginning (inline)', () => {
-    renderEmbed({
-      name: 'replay',
-      data: {id: 'abcdef1234567890'},
-      level: 'inline',
-    });
-
-    expect(screen.getByRole('link', {name: 'Replay abcdef12'})).toHaveAttribute(
-      'href',
-      '/organizations/org-slug/explore/replays/abcdef1234567890/'
-    );
-  });
-
-  it('renders a replay player preview at block level with a timestamp', async () => {
-    jest.spyOn(console, 'error').mockImplementation(() => {});
-
-    renderEmbed({
-      name: 'replay',
-      data: {
-        id: '4c1f2e3d1234567890',
-        eventTimestamp: '2026-08-25T16:37:12Z',
-      },
-    });
-
-    await waitFor(() => {
-      expect(screen.getByTestId('replay-loading-placeholder')).toBeInTheDocument();
-    });
-  });
-
-  it('falls back to a link at block level without a timestamp', () => {
-    renderEmbed({
-      name: 'replay',
-      data: {id: 'abcdef1234567890'},
-    });
-
-    expect(screen.getByRole('link', {name: 'Replay abcdef12'})).toHaveAttribute(
-      'href',
-      '/organizations/org-slug/explore/replays/abcdef1234567890/'
-    );
-  });
-
-  describe('alert', () => {
-    it('points a metric alert at its detector', () => {
-      expect(
-        hrefFor('alert', 'Checkout latency', {
-          id: '4521',
-          kind: 'metric',
-          name: 'Checkout latency',
-        })
-      ).toBe('/organizations/org-slug/monitors/4521/');
-    });
-
-    it('points an issue alert at its automation', () => {
-      expect(hrefFor('alert', 'Alert 881', {id: '881', kind: 'issue'})).toBe(
-        '/organizations/org-slug/monitors/alerts/881/'
-      );
-    });
-
-    it('falls back to an id-based label when the API name is missing', () => {
-      renderEmbed({name: 'alert', data: {id: '4521', kind: 'metric'}});
-      expect(screen.getByRole('link', {name: 'Alert 4521'})).toBeInTheDocument();
-    });
-  });
-
-  it('links a monitor to its detector detail page', () => {
-    expect(hrefFor('monitor', 'nightly-sync', {id: '9931', name: 'nightly-sync'})).toBe(
-      '/organizations/org-slug/monitors/9931/'
-    );
-  });
-
-  it('links a saved issue view to the view route', () => {
-    expect(hrefFor('savedIssueView', 'Unresolved', {id: '77', name: 'Unresolved'})).toBe(
-      '/organizations/org-slug/issues/views/77/'
-    );
-  });
-
-  it.each([
-    ['spans', '/organizations/org-slug/explore/traces/?id=312'],
-    ['logs', '/organizations/org-slug/explore/logs/?id=312'],
-    ['metrics', '/organizations/org-slug/explore/metrics/?id=312'],
-    ['replays', '/organizations/org-slug/explore/replays/?id=312'],
-  ])('opens a saved %s query on its own explore surface', (dataset, expected) => {
-    expect(hrefFor('savedQuery', 'Saved query 312', {id: '312', dataset})).toBe(expected);
-  });
-
-  describe('trace', () => {
-    const traceId = 'a1b2c3d4e5f678901234567890abcdef';
-
-    it('converts the ISO timestamp to unix seconds for the waterfall', () => {
-      const href = hrefFor('trace', 'Trace a1b2c3d4', {
-        traceId,
-        timestamp: '2026-08-25T16:37:12Z',
-      });
-
-      expect(href).toContain(`/explore/traces/trace/${traceId}/`);
-      expect(href).toContain(`timestamp=${Date.parse('2026-08-25T16:37:12Z') / 1000}`);
-    });
-
-    it('focuses a span when one is given', () => {
-      expect(hrefFor('trace', 'Trace a1b2c3d4', {traceId, spanId: 'abc123'})).toContain(
-        'node=span-abc123'
-      );
-    });
-
-    it('omits trace query params that were not provided', () => {
-      expect(hrefFor('trace', 'Trace a1b2c3d4', {traceId})).toBe(
-        `/organizations/org-slug/explore/traces/trace/${traceId}/`
-      );
-    });
-  });
-
-  it('links a profile to its flamegraph', () => {
-    expect(
-      hrefFor('profile', 'Profile 7f3c2b1a', {
-        projectSlug: 'javascript',
-        profileId: '7f3c2b1a9d8e4f60',
-      })
-    ).toBe(
-      '/organizations/org-slug/explore/profiles/profile/javascript/7f3c2b1a9d8e4f60/flamegraph/'
-    );
-  });
-
-  describe('issuesQuery', () => {
-    it('carries the search string and page filters into the issue stream', () => {
-      const href = hrefFor('issuesQuery', 'Unresolved errors', {
-        query: 'is:unresolved level:error',
-        statsPeriod: '7d',
-        projects: ['1', '2'],
-        title: 'Unresolved errors',
-      });
-
-      expect(href).toContain('/organizations/org-slug/issues/');
-      expect(href).toContain('query=is%3Aunresolved%20level%3Aerror');
-      expect(href).toContain('statsPeriod=7d');
-      expect(href).toContain('project=1');
-      expect(href).toContain('project=2');
-    });
-
-    it('uses a generic label when Seer supplies no title', () => {
-      renderEmbed({name: 'issuesQuery', data: {query: 'is:unresolved'}});
-      expect(screen.getByRole('link', {name: 'Issue search'})).toBeInTheDocument();
-    });
-  });
-
-  it('builds an errors query with columns and a sort', () => {
-    const href = hrefFor('errorsQuery', 'Checkout errors', {
-      query: 'event.type:error',
-      fields: ['title', 'count()'],
-      sort: '-count',
-      statsPeriod: '24h',
-      title: 'Checkout errors',
-    });
-
-    expect(href).toContain('/explore/discover/results/');
-    expect(href).toContain('query=event.type%3Aerror');
-    expect(href).toContain('field=title');
-    expect(href).toContain('field=count%28%29');
-    expect(href).toContain('sort=-count');
-  });
-
-  describe('spansQuery', () => {
-    it('builds a samples-mode query', () => {
-      const href = hrefFor('spansQuery', 'Span search', {
-        query: 'span.op:http.client',
-        mode: 'samples',
-        sort: '-span.duration',
-        statsPeriod: '24h',
-      });
-
-      expect(href).toContain('/organizations/org-slug/explore/traces/');
-      expect(href).toContain('mode=samples');
-      expect(href).toContain('query=span.op%3Ahttp.client');
-      expect(href).toContain('sort=-span.duration');
-    });
-
-    it('encodes group-bys and aggregates for aggregate mode', () => {
-      const href = hrefFor('spansQuery', 'p95 by op', {
-        mode: 'aggregate',
-        groupBy: ['span.op'],
-        yAxes: ['p95(span.duration)'],
-        title: 'p95 by op',
-      });
-
-      expect(href).toContain('mode=aggregate');
-
-      const params = new URL(href, 'https://sentry.io').searchParams;
-      expect(params.getAll('aggregateField').map(field => JSON.parse(field))).toEqual([
-        {groupBy: 'span.op'},
-        {yAxes: ['p95(span.duration)']},
-      ]);
-    });
-
-    it('coerces numeric project IDs so agent payloads still render', () => {
-      const href = hrefFor('spansQuery', 'Issue Pageloads (Last 30 Days)', {
-        mode: 'aggregate',
-        query: 'span.op:pageload transaction:*issues*',
-        groupBy: ['transaction'],
-        yAxes: ['count()'],
-        statsPeriod: '30d',
-        projects: [11276],
-        title: 'Issue Pageloads (Last 30 Days)',
-      });
-
-      expect(href).toContain('/organizations/org-slug/explore/traces/');
-      expect(href).toContain('project=11276');
-      expect(href).toContain('statsPeriod=30d');
-      expect(href).toContain('mode=aggregate');
-    });
-  });
-
-  it('builds a logs query using the logs-prefixed params', () => {
-    const href = hrefFor('logsQuery', 'Error logs', {
-      query: 'severity:error',
-      mode: 'samples',
-      statsPeriod: '24h',
-      title: 'Error logs',
-    });
-
-    expect(href).toContain('/organizations/org-slug/explore/logs/');
-    expect(href).toContain('logsQuery=severity%3Aerror');
-    expect(href).toContain('mode=samples');
-  });
-
-  it('builds a replays query', () => {
-    const href = hrefFor('replaysQuery', 'Replay search', {
-      query: 'count_rage_clicks:>0',
-      statsPeriod: '7d',
-    });
-
-    expect(href).toContain('/organizations/org-slug/explore/replays/');
-    expect(href).toContain('query=count_rage_clicks%3A%3E0');
-    expect(href).toContain('statsPeriod=7d');
-  });
-
-  describe('metricsQuery', () => {
-    it('encodes the metric as a single JSON param', () => {
-      const href = hrefFor('metricsQuery', 'checkout.latency', {
-        name: 'checkout.latency',
-        type: 'distribution',
-        unit: 'millisecond',
-        mode: 'aggregate',
-        yAxes: ['p95(value)'],
-      });
-
-      expect(href).toContain('/organizations/org-slug/explore/metrics/');
-
-      const metric = new URL(href, 'https://sentry.io').searchParams.get('metric');
-      expect(JSON.parse(metric ?? '{}')).toEqual({
-        metric: {name: 'checkout.latency', type: 'distribution', unit: 'millisecond'},
-        query: '',
-        aggregateFields: [{yAxes: ['p95(value)']}],
-        mode: 'aggregate',
-      });
-    });
-
-    it('charts a default aggregate so the query stays decodable', () => {
-      const href = hrefFor('metricsQuery', 'checkout.latency', {
-        name: 'checkout.latency',
-        type: 'distribution',
-      });
-
-      const metric = new URL(href, 'https://sentry.io').searchParams.get('metric');
-      expect(JSON.parse(metric ?? '{}').aggregateFields).toEqual([
-        {yAxes: ['sum(value)']},
-      ]);
-    });
-  });
-});
